ai-service/requirements.txt
# FastAPI - 웹 프레임워크
fastapi==0.115.0
uvicorn[standard]==0.30.6

# LangChain - AI 체인 구성
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
langchain-core==0.3.6

# OpenAI
openai==1.51.0

# Vector Store & Embeddings
chromadb==0.5.11
sentence-transformers==3.1.1

# Database
pymysql==1.1.1
sqlalchemy==2.0.35

# Document Processing
pypdf==5.0.1
python-docx==1.1.2
python-multipart==0.0.12

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2

# CORS
fastapi-cors==0.0.6